requirements.txt
notion-client>=2.2.1
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
httpx>=0.25.0
lxml>=4.9.0
playwright>=1.40.0
openai>=1.0.0
python-docx>=1.0.0
streamlit>=1.28.0

